Refactor: move MembersList to common
