+++ src/components/common/Members/MembersList/MemberInfo/MemberInfo.css
