+++ src/components/common/Members/MembersList/MemberInfo/MemberInfo.tsx
